# Set number of threads to 1