Before:
```
Server parameters: num_experts=16, num_handlers=64, max_batch_size=8192, expert_cls=ffn, hid_dim=1024, device=cuda
Client parameters: num_clients=128, num_batches_per_client=16, batch_size=2048, backprop=True
Results:
        Server startup took 6.825 s. (2.159 s. experts + 4.665 s. networking)
        Processed 4194304 examples in 160.663
        Throughput for forward + backward passes: 26106.234 samples / s.
        Benchmarking took 167.689 s.
```
```
Server parameters: num_experts=4, num_handlers=64, max_batch_size=8192, expert_cls=ffn, hid_dim=1024, device=cuda
Client parameters: num_clients=128, num_batches_per_client=64, batch_size=32, backprop=False
Results:
        Server startup took 2.998 s. (0.549 s. experts + 2.449 s. networking)
        Processed 262144 examples in 9.480
        Throughput for forward passes: 27652.515 samples / s.
        Benchmarking took 12.663 s.
```
After:
```
Server parameters: num_experts=16, num_handlers=64, max_batch_size=8192, expert_cls=ffn, hid_dim=1024, device=cuda
Client parameters: num_clients=128, num_batches_per_client=16, batch_size=2048, backprop=True
Results:
        Server startup took 6.827 s. (2.157 s. experts + 4.671 s. networking)
        Processed 4194304 examples in 96.610
        Throughput for forward + backward passes: 43414.985 samples / s.
        Benchmarking took 103.814 s.
```
```
Server parameters: num_experts=4, num_handlers=64, max_batch_size=8192, expert_cls=ffn, hid_dim=1024, device=cuda
Client parameters: num_clients=128, num_batches_per_client=64, batch_size=32, backprop=False
Results:
        Server startup took 2.996 s. (0.551 s. experts + 2.445 s. networking)
        Processed 262144 examples in 9.483
        Throughput for forward passes: 27642.687 samples / s.
        Benchmarking took 12.692 s.
```